chore: move most of the UI into the shell component
